Add first default paths for configuration files
This commit add a server, that should provide a access to the
configuration for TEP. Currently the configuration_server only searches
for the configuration file. It therefore searches the default
directories and gives access to the first found file.

Further development should include the old-style config file directory
on linux systems, normally located in ~/.tep/config.ini .