Add command-line flag -config_file
To specify a config file at startup, this adds a routine to check for a
command-line flag. The flag, which is used is config_flag which is to be
followed by a string, that designates the config file. If multiple of
these flags a given, only the first is used. If no flag is specified,
the atom no_path is used. The string is given to tep_sup:start_link/1
for further use.